+++ include/wx/pdfdoc_version.h
@@ -11,9 +11,9 @@
 #define _PDFDOC_VERSION_H_
 
 #define PDFDOC_MAJOR_VERSION      1
-#define PDFDOC_MINOR_VERSION      0
-#define PDFDOC_RELEASE_NUMBER     3
+#define PDFDOC_MINOR_VERSION      1
+#define PDFDOC_RELEASE_NUMBER     0
 #define PDFDOC_SUBRELEASE_NUMBER  0
-#define PDFDOC_VERSION_STRING     "wxPdfDocument 1.0.3"
+#define PDFDOC_VERSION_STRING     "wxPdfDocument 1.1.0"
 
 #endif // _PDFDOC_VERSION_H_