+++ src/pdfdc.cpp
@@ -1084,7 +1084,7 @@ wxPdfDCImpl::DoDrawRotatedText(const wxString& text, wxCoord x, wxCoord y, doubl
     }
     wxBrush previousBrush = GetBrush();
     SetBrush(wxBrush(m_textBackgroundColour));
-    SetupBrush();
+    SetupBrush(true);
     SetupAlpha();
     // draw rectangle line by line
     for (size_t lineNum = 0; lineNum < lines.size(); lineNum++)
@@ -1493,14 +1493,14 @@ wxPdfDCImpl::MustSetCurrentBrush(const wxBrush& currentBrush) const
 }
 
 void
-wxPdfDCImpl::SetupPen()
+wxPdfDCImpl::SetupPen(bool force)
 {
   wxCHECK_RET(m_pdfDocument, wxS("Invalid PDF DC"));
   // pen
   const wxPen& curPen = GetPen();
   if (curPen != wxNullPen)
   {
-    if (MustSetCurrentPen(curPen))
+    if (force || MustSetCurrentPen(curPen))
     {
       wxPdfLineStyle style = m_pdfDocument->GetLineStyle();
       wxPdfArrayDouble dash;
@@ -1597,14 +1597,14 @@ wxPdfDCImpl::SetupPen()
 }
 
 void
-wxPdfDCImpl::SetupBrush()
+wxPdfDCImpl::SetupBrush(bool force)
 {
   wxCHECK_RET(m_pdfDocument, wxS("Invalid PDF DC"));
   // brush
   const wxBrush& curBrush = GetBrush();
   if (curBrush != wxNullBrush)
   {
-    if (MustSetCurrentBrush(curBrush))
+    if (force || MustSetCurrentBrush(curBrush))
     {
       wxColour brushColour = curBrush.GetColour();
       wxString pdfPatternName;