+++ src/pdftemplate.cpp
@@ -139,6 +139,32 @@ wxPdfDocument::EndTemplate()
 }
 
 void
+wxPdfDocument::UseTemplate(int templateId)
+{
+  if (m_page <= 0)
+  {
+    wxLogError(wxString(wxS("wxPdfDocument::UseTemplate: ")) +
+      wxString(_("You have to add a page first!")));
+    return;
+  }
+
+  wxPdfTemplate* tpl;
+  wxPdfTemplatesMap::iterator templateIter = (*m_templates).find(templateId);
+  if (templateIter != (*m_templates).end())
+  {
+    tpl = templateIter->second;
+  }
+  else
+  {
+    wxLogWarning(wxString(wxS("wxPdfDocument::UseTemplate: ")) +
+      wxString::Format(_("Template %d does not exist!"), templateId));
+    return;
+  }
+
+  UseTemplate(templateId, tpl->GetX(), tpl->GetY());
+}
+
+void
 wxPdfDocument::UseTemplate(int templateId, double x, double y, double w, double h)
 {
   if (m_page <= 0)
@@ -166,8 +192,6 @@ wxPdfDocument::UseTemplate(int templateId, double x, double y, double w, double 
     (*(m_currentTemplate->m_templates))[templateId] = tpl;
   }
 
-  if (x < 0) x = tpl->GetX();
-  if (y < 0) y = tpl->GetY();
   GetTemplateSize(templateId, w, h);
 
   double xScale = w / tpl->GetWidth();